pyproject.toml
[tool.poetry]
name = "jobparser"
version = "0.1.0"
description = ""
authors = ["vvvadim <v.vladimirov@it-park.tech>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
